feat: extract statistical functions into reddwarf-ts workspace package
Adopts pnpm workspaces and moves the polis-style representative
statement and consensus analysis functions (originally ported from
raykyri/osccai-simulation) into a standalone packages/reddwarf-ts
package importable from other projects without pulling in the app.

Key changes:
- `analyzePaintedClusters` renamed `analyzeLabeledGroups` in the
  package (neutral terminology; no painting concepts leak out)
- DuckDB decoupled: package functions accept a `VoteConnection`
  interface; the app adapter in src/lib/representative-statements.ts
  handles ensureVotesTableLoaded + getConnection injection
- src/lib/stats.ts becomes a re-export shim
- App public API is unchanged; no component updates needed
