+++ package.json
@@ -22,6 +22,7 @@
     "chromatic": "pnpm dlx chromatic --project-token=chpt_38c770d91124bfb"
   },
   "dependencies": {
+    "reddwarf-ts": "workspace:*",
     "@duckdb/duckdb-wasm": "^1.30.0",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-alert-dialog": "^1.1.15",
