+++ packages/reddwarf-ts/src/index.ts
@@ -1,0 +1,2 @@
+export * from './stats.js';
+export * from './representative-statements.js';
